+++ libs/eventstore-interconnect/src/connections-guards/index.ts
@@ -1,3 +1,4 @@
 export * from './connection-guard';
 export * from './legacy/legacy-connection-guard.service';
-export * from './next/next.service';
+export * from './next/next-connection-guard.service';
+export * from './connection-guard.constants';