+++ libs/eventstore-interconnect/src/reader/services/constants.ts
@@ -1,2 +1,1 @@
 export const SUBSCRIPTIONS = Symbol();
-export const EVENTSTORE_DB_CLIENT = Symbol();