+++ libs/eventstore-interconnect/src/reader/services/legacy-clients-connection-initializers/eventstore-client/legacy-event-store-connection-initializer.service.ts
@@ -63,7 +63,10 @@ export class LegacyEventStoreConnectionInitializerService
 
     await this.eventStoreNodeConnection.connect();
 
-    await this.connectionGuard.checkTcpConnection(
+    this.logger.log(
+      `Starting to ping connection on ${tcpEndPoint.host}:${tcpEndPoint.port}...`,
+    );
+    await this.connectionGuard.startConnectionLinkPinger(
       this.eventStoreNodeConnection,
       this.configuration.source,
     );