+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Major evlolutions of the project
 
+### 1.0.26
+- New feature : Each  $CONNECTION_LINK_CHECK_INTERVAL_IN_MS milliseconds, the connection to both source and destination will be checked. When the connection timeout, process will exit with code 1
+
 ### 1.0.25
 - Connections initialized on module init (hooked to lifecycle)
 - Events are copied exactly as they are